requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
python-dotenv==1.0.1
jinja2==3.1.3
python-multipart==0.0.9
pywebview==4.4.1
pyinstaller==6.3.0
